+++ ChangeLog.md
@@ -5,6 +5,7 @@
     [ 20141127 ]
     * Started work on loading the object texture mapping, with debug UI
     * Textiles debug UI also draws triangles “properly”
+    * Removed Sound command, added Sound Debug UI
 
     [ 20141126 ]
     * Reduced code duplication for BinaryFile/BinaryMemory